+++ lib/src/entities/entities_impl.dart
@@ -2,6 +2,7 @@ export 'impl/activity_info.dart';
 export 'impl/application_info.dart';
 export 'impl/configuration_info.dart';
 export 'impl/feature_info.dart';
+export 'impl/install_source_info_impl.dart';
 export 'impl/instrumentation_info.dart';
 export 'impl/package_info.dart';
 export 'impl/permission_info.dart';